pva gateway: review nits (plausibility tests to unmarked file, CHANGELOG dash style)
